moving CancelAndStopIntent Handlers and HelpIntent handlers to separate files
