+++ lambda/custom/handlers/General/CancelAndStopIntentHandler.js
@@ -1,21 +1,5 @@
 const { ri } = require('@jargon/alexa-skill-sdk');
 
-const HelpIntentHandler = {
-	canHandle(handlerInput) {
-		return handlerInput.requestEnvelope.request.type === 'IntentRequest' &&
-			handlerInput.requestEnvelope.request.intent.name === 'AMAZON.HelpIntent';
-	},
-	handle(handlerInput) {
-		const speechText = ri('HELP.MESSAGE');
-
-		return handlerInput.jrb
-			.speak(speechText)
-			.reprompt(speechText)
-			.withSimpleCard(ri('HELP.CARD_TITLE'), speechText)
-			.getResponse();
-	},
-};
-
 const CancelAndStopIntentHandler = {
 	async canHandle(handlerInput) {
 		const { attributesManager } = handlerInput;
@@ -45,5 +29,5 @@ const CancelAndStopIntentHandler = {
 
 module.exports = {
 	CancelAndStopIntentHandler,
-	HelpIntentHandler,
 };
+
