+++ lambda/custom/handlers/General/HelpIntentHandler.js
@@ -1,0 +1,22 @@
+const { ri } = require('@jargon/alexa-skill-sdk');
+
+const HelpIntentHandler = {
+	canHandle(handlerInput) {
+		return handlerInput.requestEnvelope.request.type === 'IntentRequest' &&
+			handlerInput.requestEnvelope.request.intent.name === 'AMAZON.HelpIntent';
+	},
+	handle(handlerInput) {
+		const speechText = ri('HELP.MESSAGE');
+
+		return handlerInput.jrb
+			.speak(speechText)
+			.reprompt(speechText)
+			.withSimpleCard(ri('HELP.CARD_TITLE'), speechText)
+			.getResponse();
+	},
+};
+
+
+module.exports = {
+	HelpIntentHandler,
+};
